Download and visualisation of incidences
The android app downloads incidences that locate near the user (in a radius of 10km regarding his position) from the server (accessing the database through the web app) and visualises the near ones (located at less than 1km of distance). The incidences are downloaded periodically each 2 minutes.
Up to 3 incidences can be visualised at the same time and will be ordered by proximity, being the nearest one the biggest icon on the screen. The user is given oral warnings about incidences.