Accelerometter-Real world transformation
Transform acelerometer values to the real world frame. We accomplish this by creating a rotational matrix on the 100th lecture of the sensor.
To test this version, the terminal must be on a stand-by position, then, press the start button and wait for the notification. Beyond this point,
all the accelerometer data will be transformed to the real world frame by the multiplication of the rotation matrix and the accelerometer values.

This update will help on accelerometer data gatering by providing acurate data, position(inclination) independant.