Detect when device is not moving
Analises the accelerometer and gyroscope data to see if the device is rotating or accelerating. If only one of them is happening, the device is moving so the class Calibrator will end its thread. However, if the device is not moving, it will compute the mean of accelerations and the mean of magnetometer values and compute a rotation matrix with them. This rotation matrix is notified to the DevicePositionChangedListener passed to the constructor of the Calibrator. From now on, all accelerometer and gyroscope data should be multiplied by this matrix to obtain the original data (this data transformed to the real world system). This calculus should be done periodically to detect if the device has changed its position, to replace the rotation matrix by the new one (that represents the change to be done for converting the current axis to the real world axes).