Rolling window support
Added a rolling window to gather the data of the sensors in a 3xwindow size/sample frequencyx3 array.
To accomplish this, a window size is calculated by a sampling frequency of a minimum of 5 (actually hardcoded) and a data-gathering time (also hardcoded). Once the window is calculated, sensor data is stored dynamically in this window, writing over the first data samples once the window is full.
Bearing in mind the simultaneous window write/read accesses by multiple threads, values are protected by locks.
Finally, the time of copying the samples for an additional AI window is measured.