FileWriter saves data each line
Before this commit, all the data that was going to be saved was stored in ram until specified to be saved. Now, each line (\n, also called return) is called to the write function and before closing the file, a flush of the buffer is done (to force the write of the cache into the non-volatile memory). Furthermore, forced portrait (vertical) orientation in the android manifest.