Implemented EnergyThreshold
After passing the SVMThreshold, the possible incidence data is analysed by the EnergyThreshold. This one uses a ButterWorth style high-pass filter (in the ContinousFilter.java class) to eliminate the continuous component of the signal. Then, the energy of the signal is calculated to compare its maximum and mean. If that division surpasses the filter, it will be analysed by the artificial intelligence.