Accumulation signal testing
The project neural network tutorial contains the source-code to create an Elman Neural Network (a recurrent neural network). This network is trained by a signal (values of a variable over time) which has the following appearance: _____/¯¯\_____ . In this supervised training, the output identifies the corresponding class of each signal instance, based on its current and previous values (because of being recurrent, it holds previous input values to network in context neurons). The class number 1 is used to identify the first instance of normal signal value (_) after de peak has been passed. All the other instances are identified with the class 0.