Datasets loaded from files
There are two folders. The train folder is where the training datasets are placed. They are all loaded to train the neural network and create a model. After training, the datasets placed at test folder are loaded to test the model. The testing feeds the input and compares the output of the trained model with the ideal (hoped) output.