AI training in files
The AI loads data to train the model while logging the training in the result/logSyso file. Then, it tests it and outputs the result in the result/testing files. AI data to train and test is loaded as specified in 9a30909c4691a25142200de7f3608e84c114bb3a