Restrictive pothole detection
When the AI is analysing the collected sensor's data, it checks if a curve has been detected or not. If it has, then it won't trigger a pothole. It will in case a curve has not been detected.
This is the first approach in pothole and curve detection because potholes and curves are much alike on Accelerometer Z axis, so we are not able to distinguish them with those threshold filters.